refactor: move cart to separate class
