+++ src/shared/api/clientApi/parsers/parseProduct.ts
@@ -3,8 +3,8 @@ import type {
   Price,
   ProductProjection,
 } from '@commercetools/platform-sdk';
-import type { AttributesData, ProductData } from './types';
-import { client } from './ClientApi';
+import type { AttributesData, ProductData } from '../types';
+import { client } from '../ClientApi';
 
 export function parseProduct(result: ProductProjection): ProductData {
   const variant = result.masterVariant;
